backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-chroma
langchain-huggingface
langchain-pinecone
pypdf
sentence-transformers
transformers
torch
sse-starlette
python-multipart
